+++ Examples/Console/I2C/example_I2C_write_read.py
@@ -81,7 +81,7 @@ async def main():
     except Exception as err:
         pywpc.printGenericError(err)
 
-    ## Disconnect network device
+    ## Disconnect USB device
     dev.disconnect()
 
     ## Release device handle